requirements.txt
transformers>=4.30.0
torch>=1.12.0
numpy>=1.21.0
PyPDF2>=2.0.0
beautifulsoup4>=4.10.0
rouge>=1.0.1
scikit-learn>=1.0.0
pyyaml>=6.0
datasets>=2.0.0